db-api/requirements.txt
grpcio==1.34.0 
grpcio-tools==1.34.0 
pymongo==3.11.2 
dnspython==2.1.0 